backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic-ai>=0.0.1
python-multipart>=0.0.6
marker-pdf>=0.2.0
httpx>=0.26.0
gitpython>=3.1.41
